pyproject.toml
[project]
name = "placement-readiness-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "google-genai>=1.47.0",
    "gunicorn>=23.0.0",
    "langchain>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.2",
    "langchain-google-genai>=3.0.0",
    "langchain-text-splitters>=1.0.0",
    "pypdf>=6.1.3",
    "python-dotenv>=1.2.1",
    "werkzeug>=3.1.3",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
